+++ eslint-bridge/src/rules/main.ts
@@ -35,7 +35,7 @@ import { rule as commaOrLogicalOrCase } from "./comma-or-logical-or-case";
 import { rule as maxUnionSize } from "./max-union-size";
 import { rule as noAccessorFieldMismatch } from "./no-accessor-field-mismatch";
 import { rule as noArrayDelete } from "./no-array-delete";
-import { rule as noDeadStore } from "./no-deadstore";
+import { rule as noDeadStore } from "./no-dead-store";
 import { rule as noDuplicateInComposite } from "./no-duplicate-in-composite";
 import { rule as noGratuitousExpressions } from "./no-gratuitous-expressions";
 import { rule as noHardcodedCredentials } from "./no-hardcoded-credentials";
@@ -105,7 +105,7 @@ ruleModules["max-union-size"] = maxUnionSize;
 ruleModules["no-duplicate-in-composite"] = noDuplicateInComposite;
 ruleModules["no-gratuitous-expressions"] = noGratuitousExpressions;
 ruleModules["no-hardcoded-credentials"] = noHardcodedCredentials;
-ruleModules["no-deadstore"] = noDeadStore;
+ruleModules["no-dead-store"] = noDeadStore;
 ruleModules["no-ignored-return"] = noIgnoredReturn;
 ruleModules["no-inconsistent-returns"] = noInconsistentReturns;
 ruleModules["no-invalid-await"] = noInvalidAwait;
