+++ eslint-bridge/tests/rules/no-dead-store.test.ts
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { rule } from "../../src/rules/no-deadstore";
+import { rule } from "../../src/rules/no-dead-store";
 import { RuleTester } from "eslint";
 import { RuleTesterTs } from "../RuleTesterTs";
